modified:   lib/connect.py 	modified:   lib/gen_html.py 	modified:   lib/web_os.py 	modified:   src/main.py 	modified:   src/web_os_run.py

+++ src/main.py
@@ -13,12 +13,13 @@ time.sleep(3)
 import lib.connect as connect
 w=connect.Wifi()
 
-if w.run_mode():
+if w.creds() and w.set_sta(1) and w.run_mode():
     print("Normal mode activated - for chance to OS-mode type in terminal:")
     print(">>>import os")
     print(">>>os.remove('run_mode.dat'")    
     import main1
-else:
+else:                
     print("OS mode activated")
     import web_os_run
-    
+    web_os_run.run(w)
+
